requirements.txt
pandas
numpy 
scikit-learn
scipy
seaborn 
matplotlib 
flask
pymongo
dagshub
dvc
mlflow
dvc-s3
python-dotenv
optuna